Refactoring
Moved all of the helpers into a single helper class for the sake of code clarity.
